Post-estimation treatment effects for an ETWFE regressions.

Post-estimation treatment effects for an ETWFE regressions.

Usage

emfx(
  object,
  type = c("simple", "group", "calendar", "event"),
  by_xvar = "auto",
  collapse = "auto",
  post_only = TRUE,
  ...
)

Arguments

object

An `etwfe` model object.

type

Character. The desired type of post-estimation aggregation.

by_xvar

Logical. Should the results account for heterogeneous treatment effects? Only relevant if the preceding `etwfe` call included a specified `xvar` argument, i.e. interacted categorical covariate. The default behaviour ("auto") is to automatically estimate heterogeneous treatment effects for each level of `xvar` if these are detected as part of the underlying `etwfe` model object. Users can override by setting to either FALSE or TRUE. See the section on Heterogeneous treatment effects below.

collapse

Logical. Collapse the data by (period by cohort) groups before calculating marginal effects? This trades off a loss in estimate accuracy (typically around the 1st or 2nd significant decimal point) for a substantial improvement in estimation time for large datasets. The default behaviour ("auto") is to automatically collapse if the original dataset has more than 500,000 rows. Users can override by setting either FALSE or TRUE. Note that collapsing by group is only valid if the preceding `etwfe` call was run with "ivar = NULL" (the default). See the section on Performance tips below.

post_only

Logical. Only keep post-treatment effects. All pre-treatment effects will be zero as a mechanical result of ETWFE's estimation setup, so the default is to drop these nuisance rows from the dataset. But you may want to keep them for presentation reasons (e.g., plotting an event-study); though be warned that this is strictly performative. This argument will only be evaluated if `type = "event"`.

...

Additional arguments passed to [`marginaleffects::marginaleffects`]. For example, you can pass `vcov = FALSE` to dramatically speed up estimation times of the main marginal effects (but at the cost of not getting any information about standard errors; see Performance tips below). Another potentially useful application is testing whether heterogeneous treatment effects (i.e. the levels of any `xvar` covariate) are equal by invoking the `hypothesis` argument, e.g. `hypothesis = "b1 = b2"`.

Value

A fixest object with fully saturated interaction effects.

A `slopes` object from the `marginaleffects` package.

Performance tips

For datasets smaller than 100k rows, `emfx` should complete quite quickly; within a few seconds or less. However, the computation time does tend to scale linearly with the size of the data, as well as the number of interactions from the original `etwfe` model. Without getting too far into the weeds, the delta method of the underlying marginal effects calculation has to estimate two prediction models for *each* coefficient in the model and then compute their standard errors. So, it's a potentially expensive operation.

However, there are two key strategies that you can use to speed things up. The first is to pass "vcov = FALSE" as an argument to `emfx`. Doing so should reduce the estimation time to less than a second, even for datasets in excess of a million rows. This approach does come at the cost of not returning any standard errors. Yet it can be useful to combine our first strategy with a second strategy, which is to invoke the "collapse = TRUE" argument. Collapsing the data by groups prior to estimating the marginal effects can yield a substantial speed increase (albeit not nearly as dramatic as turning of the vcov calculations). But we do get standard errors this time. The trade-off from collapsing the data is that we lose some accuracy in our estimated parameters. Testing suggests that this loss in accuracy tends to be relatively minor, with results equivalent to the 1st or 2nd significant decimal place (or even better). By combining these two strategies, users can very quickly see how bad the loss in accuracy is on the main marginal effects, before deciding whether to estimate with the collapsed dataset to get approximate standard errors.

Summarizing, if you are worried about the estimation time for a large dataset, try the following three-step approach:

1. Run `emfx(..., vcov = FALSE)`.

2. Run `emfx(..., vcov = FALSE, collapse = TRUE)`.

3. Compare the results from steps 1 and 2. If the main parameter estimates are similar enough, then as your final model run the following to also obtain approximate standard errors: `emfx(..., collapse = TRUE)`.

Heterogeneous treatment effects

Specifying `etwfe(..., xvar = <xvar>)` will generate interaction effects for all levels of `<xvar>` as part of the main regression model. The reason that this is useful (as opposed to a regular, non-interacted covariate in the formula RHS) is that it allows us to estimate heterogeneous treatment effects as part of the larger ETWFE framework. Specifically, we can recover heterogeneous treatment effects for each level of `<xvar>` by passing the resulting `etwfe` model object on to `emfx()`.

For example, imagine that we have a categorical variable called "age" in our dataset, with two distinct levels "adult" and "child". Running `emfx(etwfe(..., xvar = age))` will tell us how the efficacy of treatment varies across adults and children. We can then also leverage the in-built hypothesis testing infrastructure of `marginaleffects` to test whether the treatment effect is statistically different across these two age groups; see Examples below. Note the same principles carry over to categorical variables with multiple levels, or even continuous variables (although continuous variables are not as well supported yet).

See also

Examples

# We’ll use the mpdta dataset from the did package (which you’ll need to 
# install separately).

# install.packages("did")
data("mpdta", package = "did")

#
# Basic example
#

# The basic ETWFE workflow involves two steps:

# 1) Estimate the main regression model with etwfe().

mod = etwfe(
    fml  = lemp ~ lpop, # outcome ~ controls (use 0 or 1 if none)
    tvar = year,        # time variable
    gvar = first.treat, # group variable
    data = mpdta,       # dataset
    vcov = ~countyreal  # vcov adjustment (here: clustered by county)
    )
mod
#> OLS estimation, Dep. Var.: lemp
#> Observations: 2,500 
#> Fixed-effects: first.treat: 4,  year: 5
#> Varying slopes: lpop (first.treat: 4),  lpop (year: 5)
#> Standard-errors: Clustered (countyreal) 
#>                                               Estimate Std. Error   t value
#> .Dtreat:first.treat::2004:year::2004         -0.021248   0.021728 -0.977890
#> .Dtreat:first.treat::2004:year::2005         -0.081850   0.027375 -2.989963
#> .Dtreat:first.treat::2004:year::2006         -0.137870   0.030795 -4.477097
#> .Dtreat:first.treat::2004:year::2007         -0.109539   0.032322 -3.389024
#> .Dtreat:first.treat::2006:year::2006          0.002537   0.018883  0.134344
#> .Dtreat:first.treat::2006:year::2007         -0.045093   0.021987 -2.050907
#> .Dtreat:first.treat::2007:year::2007         -0.045955   0.017975 -2.556568
#> .Dtreat:first.treat::2004:year::2004:lpop_dm  0.004628   0.017584  0.263184
#> .Dtreat:first.treat::2004:year::2005:lpop_dm  0.025113   0.017904  1.402661
#> .Dtreat:first.treat::2004:year::2006:lpop_dm  0.050735   0.021070  2.407884
#> .Dtreat:first.treat::2004:year::2007:lpop_dm  0.011250   0.026617  0.422648
#> .Dtreat:first.treat::2006:year::2006:lpop_dm  0.038935   0.016472  2.363731
#> .Dtreat:first.treat::2006:year::2007:lpop_dm  0.038060   0.022477  1.693276
#> .Dtreat:first.treat::2007:year::2007:lpop_dm -0.019835   0.016198 -1.224528
#>                                                Pr(>|t|)    
#> .Dtreat:first.treat::2004:year::2004         3.2860e-01    
#> .Dtreat:first.treat::2004:year::2005         2.9279e-03 ** 
#> .Dtreat:first.treat::2004:year::2006         9.3851e-06 ***
#> .Dtreat:first.treat::2004:year::2007         7.5694e-04 ***
#> .Dtreat:first.treat::2006:year::2006         8.9318e-01    
#> .Dtreat:first.treat::2006:year::2007         4.0798e-02 *  
#> .Dtreat:first.treat::2007:year::2007         1.0866e-02 *  
#> .Dtreat:first.treat::2004:year::2004:lpop_dm 7.9252e-01    
#> .Dtreat:first.treat::2004:year::2005:lpop_dm 1.6134e-01    
#> .Dtreat:first.treat::2004:year::2006:lpop_dm 1.6407e-02 *  
#> .Dtreat:first.treat::2004:year::2007:lpop_dm 6.7273e-01    
#> .Dtreat:first.treat::2006:year::2006:lpop_dm 1.8474e-02 *  
#> .Dtreat:first.treat::2006:year::2007:lpop_dm 9.1027e-02 .  
#> .Dtreat:first.treat::2007:year::2007:lpop_dm 2.2133e-01    
#> ... 10 variables were removed because of collinearity (.Dtreat:first.treat::2006:year::2004, .Dtreat:first.treat::2006:year::2005 and 8 others [full set in $collin.var])
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.537131     Adj. R2: 0.87167 
#>                  Within R2: 8.449e-4

# 2) Recover the treatment effects of interest with emfx().

emfx(mod)                 # simple average treatment effect (default)
#> 
#>     Term                 Contrast Estimate Std. Error     z   Pr(>|z|)    2.5 %
#>  .Dtreat mean(TRUE) - mean(FALSE) -0.05063     0.0125 -4.05 5.1167e-05 -0.07513
#>    97.5 % .Dtreat
#>  -0.02613    TRUE
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high, .Dtreat, predicted, predicted_hi, predicted_lo 
#> 
emfx(mod, type = "event") # dynamic treatment effect a la an event study
#> 
#>     Term                 Contrast Estimate Std. Error      z   Pr(>|z|)
#>  .Dtreat mean(TRUE) - mean(FALSE) -0.03321    0.01337 -2.484 0.01297951
#>  .Dtreat mean(TRUE) - mean(FALSE) -0.05735    0.01715 -3.343 0.00082830
#>  .Dtreat mean(TRUE) - mean(FALSE) -0.13787    0.03079 -4.477 7.5665e-06
#>  .Dtreat mean(TRUE) - mean(FALSE) -0.10954    0.03232 -3.389 0.00070142
#>     2.5 %   97.5 % event
#>  -0.05941 -0.00701     0
#>  -0.09097 -0.02373     1
#>  -0.19823 -0.07751     2
#>  -0.17289 -0.04619     3
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high, event, predicted, predicted_hi, predicted_lo 
#> 
# Etc. Other aggregation types are "group" and "calendar"

#
# Heterogeneous treatment effects
#

# Example where we estimate heterogeneous treatment effects for counties 
# within the 8 US Great Lake states (versus all other counties). 

gls = c("IL" = 17, "IN" = 18, "MI" = 26, "MN" = 27,
        "NY" = 36, "OH" = 39, "PA" = 42, "WI" = 55)

mpdta$gls = substr(mpdta$countyreal, 1, 2) %in% gls

hmod = etwfe(
   lemp ~ lpop, tvar = year, gvar = first.treat, data = mpdta, 
   vcov = ~countyreal,
   xvar = gls           ## <= het. TEs by gls
   )

# Heterogeneous ATEs (could also specify "event", etc.) 

emfx(hmod)
#> 
#>     Term                 Contrast Estimate Std. Error       z Pr(>|z|)    2.5 %
#>  .Dtreat mean(TRUE) - mean(FALSE) -0.04334    0.02702 -1.6037  0.10879 -0.09631
#>  .Dtreat mean(TRUE) - mean(FALSE) -0.03660    0.03907 -0.9368  0.34886 -0.11318
#>    97.5 % .Dtreat   gls
#>  0.009629    TRUE FALSE
#>  0.039976    TRUE  TRUE
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high, .Dtreat, gls, predicted, predicted_hi, predicted_lo 
#> 

# To test whether the ATEs across these two groups (non-GLS vs GLS) are 
# statistically different, simply pass an appropriate "hypothesis" argument.

emfx(hmod, hypothesis = "b1 = b2")
#> 
#>   Term  Estimate Std. Error       z Pr(>|z|)   2.5 %  97.5 %
#>  b1=b2 -0.006736    0.05288 -0.1274  0.89864 -0.1104 0.09691
#> 
#> Prediction type:  response 
#> Columns: type, term, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

#
# Nonlinear model (distribution / link) families
#

# Poisson example

mpdta$emp = exp(mpdta$lemp)

etwfe(
   emp ~ lpop, tvar = year, gvar = first.treat, data = mpdta, 
   vcov = ~countyreal,
   family = "poisson"   ## <= family arg for nonlinear options
   ) |>
   emfx("event")
#> The variables '.Dtreat:first.treat::2006:year::2004', '.Dtreat:first.treat::2006:year::2005' and eight others have been removed because of collinearity (see $collin.var).
#> 
#>     Term                 Contrast Estimate Std. Error        z   Pr(>|z|)
#>  .Dtreat mean(TRUE) - mean(FALSE)  -25.350      15.90 -1.59409 0.11091631
#>  .Dtreat mean(TRUE) - mean(FALSE)    1.092      41.84  0.02609 0.97918362
#>  .Dtreat mean(TRUE) - mean(FALSE)  -75.125      22.30 -3.36942 0.00075327
#>  .Dtreat mean(TRUE) - mean(FALSE) -101.824      28.10 -3.62309 0.00029111
#>    2.5 %  97.5 % event
#>   -56.52   5.818     0
#>   -80.92  83.100     1
#>  -118.82 -31.425     2
#>  -156.91 -46.741     3
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high, event, predicted, predicted_hi, predicted_lo 
#>