Post-estimation treatment effects for an ETWFE regressions.

Post-estimation treatment effects for an ETWFE regressions.

Usage

emfx(
  object,
  type = c("simple", "group", "calendar", "event"),
  xvar = NULL,
  post_only = TRUE,
  collapse = "auto",
  ...
)

Arguments

object

An `etwfe` model object.

type

Character. The desired type of post-estimation aggregation.

xvar

Optional interacted categorical covariate for estimating heterogeneous treatment effects. In other words, allows recovery of the (marginal) treatment effect for distinct levels of `xvar`. Works with binary categorical variables (e.g. "adult" or "child"), as well as multiple values.

post_only

Logical. Only keep post-treatment effects. All pre-treatment effects will be zero as a mechanical result of ETWFE's estimation setup, so the default is to drop these nuisance rows from the dataset. But you may want to keep them for presentation reasons (e.g., plotting an event-study); though be warned that this is strictly performative. This argument will only be evaluated if `type = "event"`.

collapse

Logical. Collapse the data by (period by cohort) groups before calculating marginal effects? This trades off a loss in estimate accuracy (typically around the 1st or 2nd significant decimal point) for a substantial improvement in estimation time for large datasets. The default behaviour ("auto") is to automatically collapse if the original dataset has more than 500,000 rows. Users can override by setting either FALSE or TRUE. Note that collapsing by group is only valid if the preceding `etwfe` call was run with "ivar = NULL" (the default). See the section on Performance tips below.

...

Additional arguments passed to [`marginaleffects::marginaleffects`]. For example, you can pass `vcov = FALSE` to dramatically speed up estimation times of the main marginal effects (but at the cost of not getting any information about standard errors; see Performance tips below). Another potentially useful application is testing whether heterogeneous treatment effects (i.e. the levels of any `xvar` covariate) are equal by invoking the `hypothesis` argument, e.g. `hypothesis = "adult = child"`.

Value

A `slopes` object from the `marginaleffects` package.

Performance tips

For datasets smaller than 100k rows, `emfx` should complete quite quickly; within a few seconds or less. However, the computation time does tend to scale linearly with the size of the data, as well as the number of interactions from the original `etwfe` model. Without getting too far into the weeds, the delta method of the underlying marginal effects calculation has to estimate two prediction models for *each* coefficient in the model and then compute their standard errors. So, it's a potentially expensive operation.

However, there are two key strategies that you can use to speed things up. The first is to pass "vcov = FALSE" as an argument to `emfx`. Doing so should reduce the estimation time to less than a second, even for datasets in excess of a million rows. This approach does come at the cost of not returning any standard errors. Yet it can be useful to combine our first strategy with a second strategy, which is to invoke the "collapse = TRUE" argument. Collapsing the data by groups prior to estimating the marginal effects can yield a substantial speed increase (albeit not nearly as dramatic as turning of the vcov calculations). But we do get standard errors this time. The trade-off from collapsing the data is that we lose some accuracy in our estimated parameters. Testing suggests that this loss in accuracy tends to be relatively minor, with results equivalent to the 1st or 2nd significant decimal place (or even better). By combining these two strategies, users can very quickly see how bad the loss in accuracy is on the main marginal effects, before deciding whether to estimate with the collapsed dataset to get approximate standard errors.

Summarizing, if you are worried about the estimation time for a large dataset, try the following three-step approach:

1. Run `emfx(..., vcov = FALSE)`.

2. Run `emfx(..., vcov = FALSE, collapse = TRUE)`.

3. Compare the results from steps 1 and 2. If the main parameter estimates are similar enough, then as your final model run the following to also obtain approximate standard errors: `emfx(..., collapse = TRUE)`.

See also

Examples

# We’ll use the mpdta dataset from the did package (which you’ll need to 
# install separately).

# install.packages("did")
data("mpdta", package = "did")

# Run the estimation
mod = etwfe(
    fml  = lemp ~ lpop, # outcome ~ controls (use 0 or 1 if none)
    tvar = year,        # time variable
    gvar = first.treat, # group variable
    data = mpdta,       # dataset
    vcov = ~countyreal  # vcov adjustment (here: clustered by county)
    )
mod
#> OLS estimation, Dep. Var.: lemp
#> Observations: 2,500 
#> Fixed-effects: first.treat: 4,  year: 5
#> Varying slopes: lpop (first.treat: 4),  lpop (year: 5)
#> Standard-errors: Clustered (countyreal) 
#>                                               Estimate Std. Error   t value
#> .Dtreat:first.treat::2004:year::2004         -0.021248   0.021728 -0.977890
#> .Dtreat:first.treat::2004:year::2005         -0.081850   0.027375 -2.989963
#> .Dtreat:first.treat::2004:year::2006         -0.137870   0.030795 -4.477097
#> .Dtreat:first.treat::2004:year::2007         -0.109539   0.032322 -3.389024
#> .Dtreat:first.treat::2006:year::2006          0.002537   0.018883  0.134344
#> .Dtreat:first.treat::2006:year::2007         -0.045093   0.021987 -2.050907
#> .Dtreat:first.treat::2007:year::2007         -0.045955   0.017975 -2.556568
#> .Dtreat:first.treat::2004:year::2004:lpop_dm  0.004628   0.017584  0.263184
#> .Dtreat:first.treat::2004:year::2005:lpop_dm  0.025113   0.017904  1.402661
#> .Dtreat:first.treat::2004:year::2006:lpop_dm  0.050735   0.021070  2.407884
#> .Dtreat:first.treat::2004:year::2007:lpop_dm  0.011250   0.026617  0.422648
#> .Dtreat:first.treat::2006:year::2006:lpop_dm  0.038935   0.016472  2.363731
#> .Dtreat:first.treat::2006:year::2007:lpop_dm  0.038060   0.022477  1.693276
#> .Dtreat:first.treat::2007:year::2007:lpop_dm -0.019835   0.016198 -1.224528
#>                                                Pr(>|t|)    
#> .Dtreat:first.treat::2004:year::2004         3.2860e-01    
#> .Dtreat:first.treat::2004:year::2005         2.9279e-03 ** 
#> .Dtreat:first.treat::2004:year::2006         9.3851e-06 ***
#> .Dtreat:first.treat::2004:year::2007         7.5694e-04 ***
#> .Dtreat:first.treat::2006:year::2006         8.9318e-01    
#> .Dtreat:first.treat::2006:year::2007         4.0798e-02 *  
#> .Dtreat:first.treat::2007:year::2007         1.0866e-02 *  
#> .Dtreat:first.treat::2004:year::2004:lpop_dm 7.9252e-01    
#> .Dtreat:first.treat::2004:year::2005:lpop_dm 1.6134e-01    
#> .Dtreat:first.treat::2004:year::2006:lpop_dm 1.6407e-02 *  
#> .Dtreat:first.treat::2004:year::2007:lpop_dm 6.7273e-01    
#> .Dtreat:first.treat::2006:year::2006:lpop_dm 1.8474e-02 *  
#> .Dtreat:first.treat::2006:year::2007:lpop_dm 9.1027e-02 .  
#> .Dtreat:first.treat::2007:year::2007:lpop_dm 2.2133e-01    
#> ... 10 variables were removed because of collinearity (.Dtreat:first.treat::2006:year::2004, .Dtreat:first.treat::2006:year::2005 and 8 others [full set in $collin.var])
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 0.537131     Adj. R2: 0.87167 
#>                  Within R2: 8.449e-4

# We can recover a variety of treatment effects of interest with the 
# complementary emfx() function. For example:
emfx(mod, type = "event")
#> 
#>     Term                 Contrast Estimate Std. Error      z   Pr(>|z|)
#>  .Dtreat mean(TRUE) - mean(FALSE) -0.03321    0.01337 -2.484 0.01297951
#>  .Dtreat mean(TRUE) - mean(FALSE) -0.05735    0.01715 -3.343 0.00082830
#>  .Dtreat mean(TRUE) - mean(FALSE) -0.13787    0.03079 -4.477 7.5665e-06
#>  .Dtreat mean(TRUE) - mean(FALSE) -0.10954    0.03232 -3.389 0.00070142
#>     2.5 %   97.5 % event
#>  -0.05941 -0.00701     0
#>  -0.09097 -0.02373     1
#>  -0.19823 -0.07751     2
#>  -0.17289 -0.04619     3
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high, event, predicted, predicted_hi, predicted_lo 
#> 

# Nonlinear model families are supported through the "family" argument.
mpdta$emp = exp(mpdta$lemp)
etwfe(
   emp ~ lpop, year, first.treat, mpdta,
   family = "poisson"
   ) |>
   emfx("event")
#> The variables '.Dtreat:first.treat::2006:year::2004', '.Dtreat:first.treat::2006:year::2005' and eight others have been removed because of collinearity (see $collin.var).
#> 
#>     Term                 Contrast Estimate Std. Error        z   Pr(>|z|)
#>  .Dtreat mean(TRUE) - mean(FALSE)  -25.350      4.585  -5.5290 3.2206e-08
#>  .Dtreat mean(TRUE) - mean(FALSE)    1.092      5.963   0.1831 0.85472858
#>  .Dtreat mean(TRUE) - mean(FALSE)  -75.125     19.349  -3.8826 0.00010333
#>  .Dtreat mean(TRUE) - mean(FALSE) -101.824      4.822 -21.1160 < 2.22e-16
#>    2.5 % 97.5 % event
#>   -34.34 -16.36     0
#>   -10.60  12.78     1
#>  -113.05 -37.20     2
#>  -111.28 -92.37     3
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high, event, predicted, predicted_hi, predicted_lo 
#>